Split `ostd/util.rs`
