+++ ostd/src/arch/x86/mm/mod.rs
@@ -15,7 +15,7 @@ use crate::{
         page_table::PageTableEntryTrait,
         Paddr, PagingConstsTrait, PagingLevel, PodOnce, Vaddr, PAGE_SIZE,
     },
-    util::SameSizeAs,
+    util::marker::SameSizeAs,
     Pod,
 };
